+++ pkg/analyzer/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added `StringInterpolation.firstString` and `lastString`, to express
   explicitly  that there are always (possibly empty) strings as the first
   and the last elements of an interpolation.
+* Deprecated `ImportElement.prefixOffset`, use `prefix.nameOffset` instead.
 
 ## 1.7.0
 * Require `meta: ^1.4.0`.